Change type of SCAN cursor from int to long long

+++ src/qredisclient/scancommand.cpp
@@ -1,7 +1,7 @@
 #include "scancommand.h"
 
 
-void RedisClient::ScanCommand::setCursor(int cursor)
+void RedisClient::ScanCommand::setCursor(long long cursor)
 {
     if (cursor <= 0)
         return;
